+++ .env
@@ -3,7 +3,7 @@ PORT=4000
 AUTH_MODE=false
 JWT_SECRET_KEY=secret-key
 
-PGHOST=postgres
+PGHOST=localhost
 PGPORT=5432
 PGDATABASE=test_db
 PGUSER=test_user
