--- src/entity/board.ts
@@ -1,17 +1,0 @@
-import { Entity, BaseEntity, Column, PrimaryGeneratedColumn } from 'typeorm';
-import BoardColumn from './board-column';
-import IBoard from '../types/iboard';
-
-@Entity('boards')
-class Board extends BaseEntity implements IBoard {
-  @PrimaryGeneratedColumn('uuid')
-  id!: string;
-
-  @Column()
-  title!: string;
-
-  @Column('json', {nullable: true})
-  columns!: BoardColumn[];
-}
-
-export default Board;
