--- src/resources/login/login.repository.ts
@@ -1,5 +1,0 @@
-import User from '../../entity/user';
-
-const getUser = async (login: string): Promise<User | undefined> => User.findOne({login});
-
-export { getUser };
