--- src/resources/login/login.service.ts
@@ -1,6 +1,0 @@
-import * as loginRepo from './login.repository'
-import User from '../../entity/user';
-
-const getUser = async (login: string): Promise<User | undefined> => loginRepo.getUser(login);
-
-export { getUser };
