--- src/types/iboard.ts
@@ -1,9 +1,0 @@
-import BoardColumn from '../entity/board-column';
-
-interface IBoard {
-  id: string,
-  title: string,
-  columns: BoardColumn[]
-}
-
-export default IBoard;
