--- src/types/icolumn.ts
@@ -1,7 +1,0 @@
-interface IColumn {
-  id: string,
-  title: string,
-  order: number
-}
-
-export default IColumn;
