--- src/types/itask.ts
@@ -1,11 +1,0 @@
-interface ITask {
-  id: string,
-  title: string,
-  order: number,
-  description: string,
-  userId: string | null,
-  boardId: string,
-  columnId: string
-}
-
-export default ITask;
