--- src/types/iuser.ts
@@ -1,8 +1,0 @@
-interface IUser {
-  id: string,
-  name: string,
-  login: string,
-  password: string
-}
-
-export default IUser;
